+++ components/modal/index.js
@@ -1,5 +1,7 @@
+import DatePickModal from './DatePickModal';
 import MapModal from './MapModal';
 
 module.exports = {
+  DatePickModal,
   MapModal,
 };
